Retain only vectorization-improving changes to fillK()
Using a const variable in the end condition removed one complaint of the
auto-vectorizer. Vertical fillK() calls are now vectorized.
Unfortunately, horizontal fillK() requires the use of 'asinf()' which is
not supported by the auto-vectorizer. Currently, asinf()  consumes ~7%
of the program's CPU time, so any further improvements would have a
relatively small impact on overall performance.